backend/requirements.txt
fastapi
uvicorn[standard]
python-binance
pandas
numpy
matplotlib
apscheduler
python-dotenv
openai
requests
python-dateutil
